Unified all vector tests into a single test file
